backend/requirements.txt
mysql-connector-python
fastapi
uvicorn
PyJWT
pydantic
python-dotenv